Update SurfaceReconstruction in Studio

+++ Studio/src/Application/Data/Project.h
@@ -52,7 +52,7 @@ public:
 
   /// load point files
   bool load_point_files(std::vector<std::string> file_names, bool local);
-  bool load_points(std::vector<std::vector<itk::Point<float> > > points, bool local);
+  bool load_points(std::vector<std::vector<itk::Point<double> > > points, bool local);
 
   void set_reconstructed_present(bool b);
